requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
huggingface_hub
langchain_groq